restructured project to a multi-module build
Restructured the project into three sub-components for similarity calculations
(dali-sim), optimisation search for good parameters for the similarity measures
(dali-opt) and the webservice (dali-webservice). Technically this was necessary
since dali-opt will use groovy that depends on version von ASM conflicting to
the version already required by the jersey server. The packages name conflict of
these incompatible library versions has to be solves by 'shading' one of the ASM
versions (in this case ASM 4.0 for groovy), which is only possilbe if the
components requiring groovy live in a seperate maven module. However, this
reorganisation might also be just reasonable from a design prespective as it
will encourage seperation of concerns and decoupling of components for different
tasks.

The resources folder containing data of SPARQL data sources has been moved to
the root directory of the project. A .gitignore file has been introduced to the
project to prevent accidental commits of build and log files.